Day 5

+++ Assets/Scripts/Core/Components/Weapon/Weapon.cs
@@ -1,5 +1,6 @@
 using Components.Health;
 using Core.Components.Behavior;
+using Core.Components.Health;
 using UnityEngine;
 using Random = UnityEngine.Random;
 
@@ -8,11 +9,11 @@ namespace Components.Weapon
      public class Weapon : MonoBehaviour
      {
          [SerializeField] private HealthComponent _healthComponent;
-         private float _damage;
+         private int _damage;
          private float _chanceVampirism;
          private float _chanceCritical;
          
-         public void LoadParameters(float damage, float chanceVampirism, float chanceCritical)
+         public void LoadParameters(int damage, float chanceVampirism, float chanceCritical)
          {
              _damage = damage;
              _chanceVampirism = chanceVampirism;
@@ -28,7 +29,7 @@ namespace Components.Weapon
              return _chanceCritical > Random.Range(0f, 1f) ? _damage * 2 : _damage;
          }
 
-         private float CurrentHeal()
+         private int CurrentHeal()
          {
              return _chanceVampirism > Random.Range(0f, 1f) ? _damage / 2 : 0;
          }
